requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
evaluate>=0.4.0
peft>=0.4.0
python-dotenv>=1.0.0
huggingface-hub>=0.16.0
numpy>=1.24.0
tqdm>=4.65.0
scikit-learn>=1.2.0
pandas>=2.0.0 